Implemented improvements
