+++ src/controllers/dynakube/dynakube_controller.go
@@ -112,11 +112,6 @@ func (controller *Controller) Reconcile(ctx context.Context, request reconcile.R
 	}
 
 	oldStatus := *dynakube.Status.DeepCopy()
-	updated := controller.reconcileIstio(dynakube)
-	if updated {
-		log.Info("istio: objects updated")
-	}
-
 	err = controller.reconcileDynaKube(ctx, dynakube)
 
 	if err != nil {
@@ -142,6 +137,11 @@ func (controller *Controller) Reconcile(ctx context.Context, request reconcile.R
 		if errClient := controller.updateDynakubeStatus(ctx, dynakube); errClient != nil {
 			return reconcile.Result{}, errors.WithMessagef(errClient, "failed to update DynaKube after failure, original error: %s", err)
 		}
+	}
+
+	updated := controller.reconcileIstio(dynakube)
+	if updated {
+		log.Info("istio: objects updated")
 	}
 
 	return reconcile.Result{RequeueAfter: requeueAfter}, err
